requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
pgvector==0.3.0
psycopg2-binary==2.9.9
redis==5.0.8
celery==5.4.0
langgraph==0.2.0
langchain-xai==0.1.0
discord.py[voice]==2.4.0
websockets==13.0
pydub==0.25.1
pillow==10.4.0
python-dotenv==1.0.1
prometheus-client==0.20.0
python-socketio==5.11.0
elevenlabs==1.0.0b3
langchain==0.2.0